+++ tests/encodings.rs
@@ -59,7 +59,7 @@ mod detect {
                 let mut r = Reader::from_reader(
                     include_bytes!(concat!("documents/encoding/", $file, ".xml")).as_ref(),
                 );
-                assert_eq!(r.decoder().encoding(), UTF_8);
+                assert_eq!(r.encoding(), UTF_8);
 
                 let mut buf = Vec::new();
                 loop {
@@ -67,7 +67,7 @@ mod detect {
                         Event::Eof => break,
                         _ => {}
                     }
-                    assert_eq!(r.decoder().encoding(), $enc);
+                    assert_eq!(r.encoding(), $enc);
                     buf.clear();
                     $($break)?
                 }